Add new module for project loading stuff
